+++ R/ggbetweenstats.R
@@ -65,8 +65,11 @@
 #'   With Tukey's method, outliers are below (1st Quartile) or above (3rd
 #'   Quartile) `outlier.coef` times the Inter-Quartile Range (IQR) (Default:
 #'   `1.5`).
-#' @param mean.plotting Logical that decides whether mean is to be highlighted
-#'   and its value to be displayed (Default: `TRUE`).
+#' @param mean.plotting Logical that decides whether centrality tendency measure
+#'   is to be displayed as a point with a label (Default: `TRUE`). Function
+#'   decides which central tendency measure to show depending on the `type`
+#'   argument (**mean** for parametric, **median** for non-parametric,
+#'   **trimmed mean** for robust, and **MAP estimator** for Bayes).
 #' @param point.args A list of additional aesthetic arguments to be passed to
 #'   the `geom_point` displaying the raw data.
 #' @param violin.args A list of additional aesthetic arguments to be passed to
@@ -175,7 +178,7 @@ ggbetweenstats <- function(data,
                            tr = 0.1,
                            mean.plotting = TRUE,
                            mean.point.args = list(size = 5, color = "darkred"),
-                           mean.label.args = list(size = 3),
+                           mean.label.args = list(size = 3, nudge_x = 0.4, segment.linetype = 4),
                            notch = FALSE,
                            notchwidth = 0.5,
                            outlier.tagging = FALSE,
@@ -406,7 +409,8 @@ ggbetweenstats <- function(data,
         x = {{ x }},
         y = {{ y }},
         k = k,
-        inherit.aes = TRUE,
+        type = type,
+        tr = tr,
         sample.size.label = sample.size.label,
         mean.point.args = mean.point.args,
         mean.label.args = mean.label.args