+++ R/reexports.R
@@ -27,3 +27,11 @@ ipmisc::enframe
 #' @export
 #' @importFrom ipmisc as_tibble
 ipmisc::as_tibble
+
+#' @export
+#' @importFrom effectsize effectsize
+effectsize::effectsize
+
+#' @export
+#' @importFrom bayestestR bayesfactor
+bayestestR::bayesfactor